# protocol/swagger.yml
...
swagger: '2.0'
info:
  description: >-
    Kuberstack Installer protocol
  version: 0.0.1
  title: kuberstack-installer
  contact:
    email: daniel@arilot.com
  license:
    name: Copyright (C) Arilot - All Rights Reserved. Proprietary and confidential.
basePath: /api/installer
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
security:
  - APIKeyHeader: []
paths:
  /auth:
    get:
      tags:
        - installer
      summary: Get session token
      operationId: getSessionID
      security: []
      responses:
        "200":
          description: Operation completed, see status
          schema:
            $ref: '#/definitions/getSessionIdOKBody'
        "500":
          description: Operation error
  /aws/regions:
    get:
      tags:
        - installer
      summary: Get session token
      operationId: getRegions
      responses:
        "200":
          description: Operation completed, see status
          schema:
            $ref: '#/definitions/getRegionsOKBody'
        "401":
          $ref: '#/responses/UnauthorizedError'
        "504":
          $ref: '#/responses/AWSTimeoutError'
        "500":
          $ref: '#/responses/InternalServerError'
...
  /cluster/domains:
    get:
      tags:
        - installer
      summary: This method should return list of the domains already registered for the corresponding account.
      operationId: getDomains
      responses:
        "200":
          description: Operation completed, see status
          schema:
            $ref: '#/definitions/getDomainsOKBody'
        "401":
          $ref: '#/responses/UnauthorizedError'
        "504":
          $ref: '#/responses/AWSTimeoutError'
        "500":
          $ref: '#/responses/InternalServerError'
...
definitions:
...
  getDomainsOKBody:
    properties:
      message:
        $ref: '#/definitions/statusMessage'
      status:
        $ref: '#/definitions/statusStatus'
      domains:
        $ref: '#/definitions/stringArray'
    type: object
    x-go-gen-location: operations
...
  getRegionsOKBody:
    properties:
      regions:
        $ref: '#/definitions/stringArray'
        description: List of AWS regions
      message:
        $ref: '#/definitions/statusMessage'
      status:
        $ref: '#/definitions/statusStatus'
    type: object
    x-go-gen-location: operations
  getSessionIdOKBody:
    properties:
      message:
        $ref: '#/definitions/statusMessage'
      status:
        $ref: '#/definitions/statusStatus'
      token:
        description: Session ID to be used with other requests
        type: string
    type: object
    x-go-gen-location: operations
...
  statusMessage:
    description: Error message (empty for status `true`)
    type: string
...
  statusStatus:
    description: Operation status
    type: boolean
  stringArray:
    items:
      type: string
    type: array
...
securityDefinitions:
   APIKeyHeader:
     type: apiKey
     in: header
     name: X-API-Key
responses:
...
  UnauthorizedError:
    description: API key is missing or invalid
    headers:
      WWW_Authenticate:
        type: string
  InternalServerError:
    description: Something get wrong
  AWSTimeoutError:
    description: AWS request timeout
...